MDEV-27159 Re-design the upper level of handling DML commands
This is the first commit for the task. This patch allows to execute only
single-table and multi-table UPDATE statements using the method
Sql_cmd_dml::execute(). The code that handles DELETE and INSERT statements
has not been touched.

Moreover, these are not the final changes to handle UPDATE statements.
All tests from the main suite passed. With --ps-protocol one test from
opt_trace_security returns not the same result. This will be fixed soon.