MDEV-5816: Stored programs: validation of stored program statements
This is the prerequisite patch to move the data member
LEX::trg_table_fields to sp_head::m_trg_table_fields.

This data memeber is used for handling OLD/NEW pseudo-rows inside
trigger body and in order to be able to reparse a trigger body
this must be moved from the struct LEX to the class sp_head.