Replace log_sys.BLOCK_SIZE with log_sys.get_block_size()
log_sys.get_block_size(): Return the physical block size of the log file.
This is only implemented on Linux and Microsoft Windows for now, and for
the power-of-2 block sizes between 64 and 4096 bytes (the minimum and
maximum size of a checkpoint block). If the block size is anything else,
the traditional 512-byte size will be used via normal file system
buffering.

If the file system buffers can be bypassed, a message like the following
will be issued:

InnoDB: File system buffers for log disabled (block size=512 bytes)
InnoDB: File system buffers for log disabled (block size=4096 bytes)

This has been tested on Linux and Microsoft Windows with both sizes.

TODO: Implement a native mechanism on PMEM. At least one PMEM device
seems to report a physical_block_size of 4096 bytes, while the
interface should be 64 bytes (the size of a cache line).