MDEV-27025 insert-intention lock conflicts with waiting ORDINARY lock
The code was backported from 10.5 be8113861c2d12bb0f7048523f886cab53326d8e
commit. See that commit message for details.

Some tests in main.xa fail because the queries don't cause deadlock
anymore. As this backport is only for custom build, the tests will be
removed in the further commit. If we are going to commit it in trunk,
the tests must be fixed.