MDEV-27402 'asm goto' is not supported on Apple Xcode 9.4.1
Even though Apple Xcode is based on and similar to Clang,
it does not support the asm goto construct which was added
in commit 668a5f3d122ef3c935bf365d8084fb9713f44c60 (MDEV-26720)
to work around a compiler deficiency that results in suboptimal
code being generated for IA-32 and AMD64.

We will disable this manual optimization if __APPLE_CC__ is defined.

Similar constructs are also used in sync/srw_lock.cc, but that code
is not used on Apple, because we have not implemented a Linux futex
compatible interface on Apple macOS.

Thanks to Valerii Kravchuk for reporting and testing this.